MVVM Base code
