+++ app/src/main/java/com/nexters/yetda/android/MainContact.kt
@@ -1,0 +1,11 @@
+package com.nexters.yetda.android
+
+class MainContact {
+    interface View {
+        //
+    }
+
+    interface ViewModel {
+        fun getFirebaseSampleData()
+    }
+}
