+++ app/src/main/java/com/nexters/yetda/android/MainViewModel.kt
@@ -1,0 +1,31 @@
+package com.nexters.yetda.android
+
+import android.util.Log
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import com.google.firebase.firestore.FirebaseFirestore
+
+class MainViewModel : ViewModel(), MainContact.ViewModel {
+
+    private val TAG = "MainViewModel"
+
+    var name = MutableLiveData<String>()
+
+    override fun getFirebaseSampleData() {
+        val db = FirebaseFirestore.getInstance()
+
+        db.collection("presents")
+            .get()
+            .addOnSuccessListener { result ->
+                for (document in result) {
+                    Log.d(TAG, "${document.id} => ${document.data}")
+                    Log.d(TAG, document.data["name"].toString())
+                    name.value = document.data["name"]?.toString()
+                }
+            }
+            .addOnFailureListener { exception ->
+                Log.w(TAG, "Error getting documents.", exception)
+            }
+    }
+
+}
