+++ build.gradle
@@ -4,7 +4,7 @@ buildscript {
     ext{
         //project dependencies
         kotlin_version = '1.3.41'
-        android_gradle_version = '3.4.2'
+        android_gradle_version = '3.5.3'
         google_service_version = '4.3.2'
         fabric_version = '1.31.2'
 
@@ -12,6 +12,7 @@ buildscript {
         appcompat_version = '1.1.0'
 //        annotation_version = '1.0.2'
         constraint_layout_version = '1.1.3'
+        lifecycle_version = '2.1.0'
 
         //firebase
         firebase_analytics_version = '17.2.1'
